requirements.txt
fastapi[all]
sqlalchemy
passlib
python-dotenv
python-jose
uvicorn
